views authenication/Permission

+++ app/routers/note.py
@@ -1,5 +1,5 @@
 from typing import List
-from .. import models, schemas
+from .. import models, schemas, oauth2
 from fastapi import Response, status, HTTPException, Depends, APIRouter
 from sqlalchemy.orm import Session
 from .. database import get_db
@@ -11,17 +11,18 @@ router = APIRouter(
 
 
 @router.get("/", response_model=List[schemas.NoteResponse])
-def get_notes(db: Session = Depends(get_db)):
+def get_notes(db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     
-    notes = db.query(models.Note).all()
+    notes = db.query(models.Note).filter(models.Note.owner_id == current_user.id).all()
     
     return notes
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.NoteResponse)
-def create_notes(note : schemas.NoteCreate, db: Session = Depends(get_db)):
-
-    new_note = models.Note(**note.dict())
+def create_notes(note : schemas.NoteCreate, db: Session = Depends(get_db),
+                current_user: int = Depends(oauth2.get_current_user)):
+    
+    new_note = models.Note(owner_id=current_user.id, **note.dict())
 
     db.add(new_note)
     db.commit()
@@ -30,7 +31,7 @@ def create_notes(note : schemas.NoteCreate, db: Session = Depends(get_db)):
     return new_note
 
 @router.get("/{id}", response_model=schemas.NoteResponse)
-def get_post(id: int, db: Session = Depends(get_db)):
+def get_post(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
 
     note = db.query(models.Note).filter(models.Note.id == id).first()
 
@@ -38,10 +39,14 @@ def get_post(id: int, db: Session = Depends(get_db)):
         raise HTTPException (status_code=status.HTTP_404_NOT_FOUND,
                                 detail=f"note with id: {id} was not found")
 
+    if note.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform request action")
+
     return note
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_note(id: int, db: Session = Depends(get_db)):
+def delete_note(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
 
     note_query = db.query(models.Note).filter(models.Note.id == id)
 
@@ -50,6 +55,10 @@ def delete_note(id: int, db: Session = Depends(get_db)):
     if note == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail= f"post with id: {id} does not exist")
+
+    if note.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
     
     note_query.delete(synchronize_session=False)
     db.commit()
@@ -57,7 +66,8 @@ def delete_note(id: int, db: Session = Depends(get_db)):
     return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 @router.put("/{id}", response_model=schemas.NoteResponse)
-def update_note(id: int, updated_note: schemas.NoteCreate, db: Session = Depends(get_db)):
+def update_note(id: int, updated_note: schemas.NoteCreate, db: Session = Depends(get_db),
+                current_user: int = Depends(oauth2.get_current_user)):
 
     note_query = db.query(models.Note).filter(models.Note.id == id)
 
@@ -66,6 +76,10 @@ def update_note(id: int, updated_note: schemas.NoteCreate, db: Session = Depends
     if note == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} does not exist")
+
+    if note.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
 
     note_query.update(updated_note.dict(), synchronize_session=False)
